Refreshed pmm

+++ kernel/arch/i386/util.c
@@ -124,17 +124,17 @@ char* itoa(int32_t value, char* buffer, int base)
     return reverse(buffer, 0, i - 1);
 }
  
-void _set_bit(uint32_t *word, uint32_t bit ){
+void _set_bit(uint8_t *word, uint8_t bit ){
     *word |= (1 << bit);
     return;
 }
-void _clear_bit(uint32_t *word, uint32_t bit){
+void _clear_bit(uint8_t *word, uint8_t  bit){
     *word &= ~(1 << bit);
     return;
 }
 
 
-bool _check_bit(uint32_t * word, uint32_t bit){
+bool _check_bit(uint8_t * word, uint8_t bit){
     return *word & (1 << bit);
 }
